createbasic functions and unittests for the project
My goal with this project is to develop an end to end pipeline via the
terminal such that someone who wants to examine the quality of their articles
versus a QnA system can feed them in, and get results out, with as much
automation and simplicity as possible.

I have taken a lot of lessons from the development world I currently work in
so I am trying to get things like instructions, readmes, requirements files
, pythondocs, unittests, etc. set up right from the beginning. Not only
will this help me if I have to come back to this later, but it also helps me
provide this as a tool to others such that they can use it with minimal
assistance

I should note that the readme is yet to be completed. I know I need to add
such things like Elasticsearch installation instructions, references to the
AI2 project along with the ARC-Solvers repo (which is the QnA system I plan
to use). Additional things that will be added to the project include tools
for a command line interface, functional tests, article importing, Data
Generation (focused for ARC) and depending on the ease of it, potentially
even an ability to auto run ARC solves for ease of use. I also need to add
a config file for the use of others (for properties difficult to specify
at the commandline).