Add `__main__.py` to allow for a terminal interface add property loading
With the experience of some of my other projects at work, doing the terminal interface was familiar
and not difficult (thankfully). This gives me a framework to expand this as time goes on. I also
added functionality to import configurations from a yaml configuration file. I debated between yaml
and json but since yaml doesn't require quite as much formatting, I decided to use yaml. I shifted
some of the hard coded values to the config file and I will try to get that tested when I next have
time to work on this (hopefully tomorrow, trying to put in ~an hour a day on this during the week).

I also added another unit/integration test, this time to test the loading of config yaml files.
Pretty simple stuff but it will help to have that later on. Doing my best to conform to the same
stringent work requirements surrounding python coding. Both for the best practices that they are
but so that others that might see this can understand it better. With the terminal interface now
open for use, and config files ready to use, I plan on working with the article parsing tomorrow
and digging into the format that I was given from a professor I'm working with. They are using
jsonl, which I'm kind of familiar with, part of me wishes it was just json though. Not sure how
best to handle the variety of formats people might use to provide articles outside of just me,
I'll probably look to that professor for some thoughts on that for ease of use. Ideally I would
avoid having an import per file type (e.g. .txt and .json and .jsonl, so on and so forth) but
we'll see what happens as I want to make this tool available for use by the group.

Tests passing, coverage raised but not 100% just yet. Will need to do some mocking to technically
reach that peak. Ideally once this system is set up I can add some larger integration and functional
tests to ensure it works properly at a larger scale.