Add in file loading for JSONL article files as specified by professor given files
This commit includes handling for the loading in of JSONL files containing JSON formatted articles.
I have put in some logic to make the usage a little easier, including the appending of the `.jsonl`
extension if not present, and the importing of a series of files rather than just a singular `.jsonl`
file. This may need to be amended if I wish to import the articles in separate objects (making an
list of lists) but I will talk with the professor about that to see what they think. With the set
of articles I have, memory does not appear to be a major hurdle to overcome, as I could import all
2112 articles and keep them in memory. I continue to expand coverage, and will have the Elasticsearch
functions added to the list of unit tested items once I've gotten a better handle on how that should
be laid out.