Alters `__main__.py` to conform to pep8 standards
Very small commit for today. Working hard on getting parsing set up for questions, but I'm not sure how
reusable the code will be. I'm going to be using the TQA dataset questions for this test by the
suggestion of the professor I work with. I'm using a particular, small set of questions, so I'm making
a small script to do that. If it is likely that other questions are needed, not just these, I'll come
back and make the script (which I'm not committing for now) will be a full fleged set of functions that
can be used, etc.