Extract title from articles to use as index name
This is another small commit with some shennanigans going on in the background. First, Elasticsearch
API is weird and confusing. I knew this from work stuff but I'm just glad I knew what I did. I was
using the `create` operation but the `index` operation is what I swapped to. The difference is that if
you use `create` it tries to create a new file and give it an id, no bueno. However, with the `index`
operation, it managed to add the document or update if it already existed. Not quite sure why this was
necessary in the first place, and I should look into it more. Either way, managed to figure it out.

Aside from that, to make sure I could actually insert the articles into the DB I needed to include the
index name from the jsonl file. I added that and adjusted the tests accordingly. I still need to do
some extra work on the unit tests and add mocking to test the DB stuff so I can get to 100% coverage.
Until then, my coverage is only a C+. Progress of the mind more than the code.