[MAJOR] rename module, update config, add requirement versions
Been a couple weeks or so since I've made a last commit. Suffice to say, hard at work, just not on
things I could commit until today. I have managed to write up a script to parse through a particular
question file from the TQA question dataset. I'll be adding the code for that hopefully soon, though I
need to reorient with my professor to get a good idea of next steps.

This commit was mostly about some quality of life related stuff. Firstly, I noticed that running this
repeatedly, sucked. So I created a fallback for the `-f` configuration to look for a property in the
config file. So, priority to terminal command options, then to the config file, failing if neither
has a proper data filepath. This will allow me to test stuff more rapidly, which I'm happy about. In
addition I have renamed the module from pipeline to archiver. This more accurately represents what the
module does, and as such, could be run on its own. My ultimate plan would be to have something that
could end to end run the archiver and several other functions in succession. I have updated `readme.md`
with the configuration and updated running instructions.

Working with Elasticsearch I encountered a strange issue where my "index" commands were being rejected
due to some "read-only" property. Initially I thought this was because of a setting in Elasticsearch,
but further Googling determined that it was actually me not having enough disk space available. After
cleaning up my disk, and re-running, I had no issues. Another oddity I noticed, but is intermittent,
is that if I rerun the archiver with the same files, it might double (but no more if rerun past that)
the size of the index. I found no evidence that it was just duplicating information beyond the index
size, as the number of hits didn't seem to change. Since I can't really reproduce it, I'm keeping it
in mind if stuff comes up later.

Quick note, in my examination of the Elasticsearch documentation, I found out that the mapping
"doc type" is a deprecated function piece. I have removed it both from my configuration file and the
code to support this. I need to make some notes about what version of Elasticsearch I expect.

I raised test coverage to 100%! I had to make some alterations. I moved the Elasticsearch constructor
call to `__main__.py` as well as the importing of `elasticsearch.helpers.bulk`. I really tried to find
a way to mock those libraries but following the online documentation of the mocking library didn't help
nor did looking for my issues in particular. With this in mind, I think this solution is workable for
now and I can double back and fix this later when its not late in the evening and I have more patience.

Another important thing to note, I have added specific versions for the different libraries I use.
While earlier/later versions are likely to work, from my recent and horrible experience with another
library that did not version (they used specific git commits *shudder*) I never wish to inflict that
pain upon anyone else that would use my code. This way, the version is specified, so everyone knows.