refactor directory structure begin to add question_extractor
Back at it again, now with 2x the purpose! In all seriousness, I think I have some solid direction.
With that work I did around the article extraction and saving to Elasticsearch I gave myself a pretty
good headstart. The next step is to be able to generate questions and challenge test files to match.
I've got a pretty good understanding of how that should work.

This commit is laying some groundwork, it's not perfect and I think it is going to break some stuff,
but the code wasn't complete as it was. Next time around I'll button up the jsonl generation then
refactor some of the code. My ultimate intention is to make something able to run the article extraction
question extraction, and the ARC Solver all with me pressing one button! I think it is possible. I know
how to make python scripts and I know I can do it!