Update readMe and do some Idiot-proofing
Only had a little amount of time today. I updated some of the readme, more is needed. In addition, I put
in a small handler to catch if someone is doing something stupid, like changing the test_set_directory
to one of the important directories. I'll be adding to this more tomorrow.