refactor unit/integration tests updates readme.md
I took some additional time and updated the readme a little more. It should be pretty comprehensive as
how the project stands right now. In addition, I have also updated the unit tests. I'm at 61% coverage
and plan on increasing that to 100% soon. Much of the functions I will be writing will need either
heavy mocking or to be full fledged integration tests. Outside of Elasticsearch and anything that runs
something on the commandline, I want to try to avoid doing mocking, so much will be more of integration
tests. That said, come the actual runner of the ARC Solver, that's going to mocking 6 ways till sunday.