100 coverage, moved constant strings into `constants.py`
AT LONG LAST 100% COVERAGE! Streamers, Balloons, and champagne for everyone! I've managed to fully test
the core pieces of functionality of steps 1 and 2. The next step is to use the output from those two
steps to do step 3: running the ARC-Solver. This next step....may be problematic. I had enough trouble
getting ARC-Solvers to run on the commandline, now I'm throwing in the monkey wrench of automation.
I'll need to do some initial test with the output of the commands and potentially make some small
adjustments, once that's done, I may even need to do some alterations to the ARC-Solver repo itself,
as I'm not sure that I'm able to change the Elasticsearch repo on the fly, which will be a pain. If I
can't I will add some sort of command line option to allow me to do it via commandline. It would also
be useful to see a tally of which questions, not how many, were answered correctly or incorrectly.

Steps projected for step 3: clean ARC-solver data directory of all except dev and training sets, copy
the test set over, as a renamed file, run the pipeline for that test set given the new elasticsearch
directory, and tally the output. Thinking about it, I may need to make some additional modifications.
The end goal here is to look at the article files and which one has the best articles overall.

I think the metrics I'll be most interested in collecting will be:
* gross article set correct answers
* average article set correct answers
* some sort of std dev or variance for the correct answers
* a ranking of the article methods based average article set correct answers

More to come soon!