add in command to run ARC-Solver via benchmark and extract results from the run
Lot of progress here, really happy it went this smoothly. I thought it would be a royal pain to get
python to be able to communicate with ARC-Solver, but I figured it out. See ARC-Solver by default
uses Conda as an environment keeper. So I used `subprocess.run` in conjunction with the
`conda run` command and I was able to force conda to run the ARC-Solver script in the conda environment
that had been set up already. This is huge for me! To top it off, I was able to get the results from
the run of the script and get the number of correct and incorrect answered questions. We'll have to
see how things go, but maybe 10 more hours on this project will let me get to a place I can run the
benchmark and be confident in how it works. Definitely gonna need to unit test this, that'll be
difficult but doable.

Added some additional configs to benchmarkConfig.yaml, not all things make sense to enter via the
terminal, I'm considering removing those entirely so it is simple how to configure it, I'll think about
that some more. The next step is to set up the automation of running the index. I already have the
reset and copy test set functions all set up. There's some additional work I need to do to get
the articles associated with their files... that'll be a little sucky but I'll make it work.

Lacking a lot of coverage for arc_runner.py, I'll add it soon.