update unit tests and add file-index dict
I fixed some unit tests that weren't in order from some changes from yesterday and some that were broken
today. In addition, I spent around 30 minutes trying to think of the best way to associate indices
in Elasticsearch with files to be able to aggregate results for article files. I landed on a method
that returns a dict after creating inserting the articles, that should allow me to run the benchmark
then associate index to file.