fixes multi-article processing and adds checkpointing
Big update here, at least in terms of time investment, haven't updated the repo in a bit,
been sick, busy, crazy stuff with the coronavirus. This update theoretically gets the main
code processing in line, the running of the ARC-Solver. In addition, it adds the checkpointing
mechanism to make it so I can run for a while leave it, come back, etc. Should make this a lot
easier.

Learned something interesting. Firstly, despite some documentation I was reading, using `a+`
as the method of opening a file, doesn't start the read at the beginning. I had to open
a read, then close after reading stuff, then open an append.

With this done, I think I can start setting things up to process the results now and get some
numbers. I want to add unit tests, documentation, and then some basic result findings along
with some tech to save it to a file rather than pure output.