Adds results analysis and file writing
To make sure the checkpoints are used, I added some minor amounts of analysis, just tallying the
counts of correct, incorrect, and unanswered entries. In addition, I write the tallied entries
and overall results to files. This plus some additional logic to prevent redoing work will help
make it easier to do further analysis on the results in the code.