Adding Elasticsearch functionality to skip re-insertion
To save time on repeat runs I am adding a change to prevent the need for reinsertion of entries
that already exist in Elasticsearch. I didn't do this before because I was worried that people
would need to overwite entries, but it's just as simple to delete the offending indicies and
reinsert. I could always later make that configurable. I adjusted the unit tests to match.

I next need to trouble shoot some issues I was encountering running the pipeline then add some
additional analysis to this.