Add TQA and ARC corpi to benchmark, fix random baseline, add question set analysis
This is a huge update, I haven't committed much lately because the code wasn't usually in
a fully working state. To my knowledge, though I need to add unit tests and documentation,
the code is working.

The first major piece of this was getting the random baseline to work correctly. I spent
a few days just trying to figure out why the question counts and the results weren't
matching up. I figured out that some indices correspond to multiple question sets, and
ultimately I was able to resolve the issue.

The second major effort was to run the ARC Benchmark on the ARC Corpus itself packaged
with the ARC-Solver. This was a bit harder to integrate because the ARC-Solver Corpus
doesn't cleanly divide into different articles, hence a large index just for the one
article was created. I used the existing process as part of the ARC-Solver project
to create that. I also discovered that inserting those lines into elasticsearch
should be done before other articles are stored in, else it will through errors related
to http traffic size.

The third major effort was around getting the TQA dataset inserted into elasticsearch
and evaluated similar to other articles. I have a basic understanding of the TQA dataset
and managed to get something set up that can be tweaked later if necessary.

The last effort I took was to get some analysis on the question sets themselves so that
I would be able to look at which question sets were harder than others. At the moment
it looks at average correct, incorrect, and unanswered, also by percentage for each and
standard deviation.

My next step will be to look to other suggested metrics that I should collect my professor
recommended to me. In addition, once I get those done, getting unit/integration tests
as well as a fully updated readme would be good.