Add text to TQA indices for non-topic but useful text
To make sure that the TQA article set gets a fair chance at answering the questions, I
have updated the logic to pull from more than just the "topics" subsection of the TQA
dataset and look at others like diagramAnnotations, instructionalDiagram, and
adjunctTopics. Some of the text may not be helpful but it is fair training data so I will
include it in the TQA indices. I hope it will positively impact the TQA article scores
since the TQA articles should be the baseline gold standard here.