Updates readme, adds documentation, additional individual question analysis
This update is mostly stuff to help others look at my code. I did some additional function
documentation that I was putting off, and I did a big update to the `readme.md`. I had been
putting it off and I finally got it in a better state, I'll need to re-read it later on
to make sure it all applies but I got instruction steps, I think, for pretty much
everything that needs to be done as part of setup to date. I believe the steps to be
straightforward and clear but it may take a different set of eyes to truly determine that.