Refines Question Analysis logic, adds 2 unit tests, includes patch
Last commit I forgot to add the actual .patch file, silly me. This commit actually includes it.
Besides this, I have refined the logic around Question set analysis. I renamed some of the
existing values to be easier to understand, and made the individual question analysis more
formal. Now disagreement is a calculated value rather than a threshold, so that someone could
look at it on their own. I've also added a couple unit tests, I'm going to try to go back
through to get this project in tip-top shape.

There's a bug that's been, well, bugging me having to do with the actual running of the
benchmark. When I had to rerun it for a 3rd time, it required more work because letting it run
for extended periods caused it to get bogus results from the benchmark that didn't happen when
I reran it after some tweaking of the checkpoint data. It's unfortunate this is happening at
all, frankly, not quite sure why this is a problem now, but if I had to guess, it might be
that pulling the info from elasticsearch this much with that much data in it could be too much
for my computer. I need to put some logic in to filter out the bad results and let it rerun
some number of times, and quit if it doesn't fix it by then.