Adding additional unit tests for results_analysis.py
With this commit, only 1 more function within results_analysis.py needs attention, that being
the old test for analyze_results. It needs to be updated for the inclusion of the baseline which
should be relatively simple. Once this is done, I need to go back and adjust broken unit tests
and then also fix the issue with long term runs of ARC-Solver.