Fixed broken unit tests, increased overall coverage
To make sure the whole project is covered by unit or integration tests, I fixed the broken
unit tests and then started to add to the unit tests to cover for missing tests. At the
moment I need to add some testing to file-utils.py and arc_runner.py. Overall coverage is
94%