altering scripts and fixing broken tests
Trying to finish up this repository for others to use. I decided that I'm going to not do
coverage for the separate scripts since they are incidental, not necessary for the basic code
and were intended for me to get additional insight, not necessary for the metrics.

With the tests fixed I need to expand coverage, re-go through my read me, and maybe finalize
the handling of the Elasticsearch overloading bug. Added some code to gracefully handle that