Adding instructions and examples for creating articles
Added an example specifically on the setup of article files and included some more in-depth
instructions and clarifications. I then linked these two files to the main readme file.

All I need to do is some small code changes/verifications to see if TQA is required by EXAM
and remove that need if it does.